v2/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=1.4.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
